DSPBRIDGE: Implemented Trampoline support for dynamic loader
"Trampolines" utilize memory space containing a local jump
table to hit external locations. Instead of the "near"
function call directly calling the external location it
"calls" a trampoline entry.

A "near" function call requires the target to be within a
signed, 21-bit offset from the current PC so all calls have
to be within 4MB of the current execution point.

The trampoline entry (location in memory) contains an
unconditional long branch to the external address to be
called. When the called function returns, the PC is returned
to the original calling location since that is where the
link register points, as the trampoline utilizes a branch
instruction.

Utilizing this method removes the 4MB location restriction
as well since the long branch can be done within all of DSP
addressable space.

This doesn't require any change on the dsp side, it is an
enhancement for newer DSP side binaries.

Signed-off-by: Fernando Guzman Lugo <x0095840@ti.com>