omap3isp: Remove redundant error handling
There is a check in isp_buf_process() which checks whether the CCDC or the
resizer were busy after setting the new buffer address. This is never the
case since CCDC is disabled before setting that address and the resizer is
used in oneshot mode. This patch removes the check.

Signed-off-by: Sakari Ailus <sakari.ailus@maxwell.research.nokia.com>