omap3isp: isp.c: check and clear CSIB receiver interrupts while wait_hs_vs
For some sensors (like stingray), after a _restart_ from
sw standby state, starting couple of frames are erroneous.
(Stingray datasheet: "When sensor _restarts_, Normal image
can be get 2 frames after")

So while we wait for HS_VS,check and clear the CSIB error
interrupts, if any. That way we ignore the CCP2 errors after a
sensor restart(stream-on -> stream-off -> stream-on) for
the first two frames.

Fixes: NB#122155 - Still image capture always generates CCP2 errors

Signed-off-by: Vimarsh Zutshi <vimarsh.zutshi@nokia.com>