et8ek8: at power on, first enable clocks and after that release reset
According to et8ek8 datasheet, when disabling camera power, sensor
extclk should be first stopped and after that the sensor put into
reset by setting XSHUTDOWN to low. However, this causes sometimes
corrupted images (even though the sensor is resetted).

This patchs modifies the power down/up sequence so that extclk is
always running when camera is not in reset state (XSHUTDOWN high):
- at power off, first set XSHUTDOWN low, then stop extclk
- at power on, first enable extclk, then set XSHUTDOWN high

Fixes: NB#125157 - Color channel missing from photos
Fixes: NB#121907 - Camera produces black pictures
Fixes: NB#125170 - Some images have wrong colours
Fixes: NB#120051 - Vertical stripes observed in both the camera view finder and resulting images
Fixes: NB#115833 - still image capture hangs in VIDIOC_DQBUF ioctl call
Fixes: NB#111089 - Image's color scheme is not correct

Signed-off-by: Tuukka Toivonen <tuukka.o.toivonen@nokia.com>