adp1653: map V4L2_POWER_STANDBY to V4L2_POWER_ON
To get synchronization with firing the flash and capturing
an image, user space should first queue buffers, then strobe
the flash, and last initiate streaming. With the new
power management model, when streaming is enabled, omap34xxcam
requests V4L2_POWER_ON mode when the flash is in STANDBY
mode. This transition resetted the flash and the first frame
was not properly exposed.

Signed-off-by: Tuukka Toivonen <tuukka.o.toivonen@nokia.com>