ZOOM2: Add OMAP3 power constraints for Sony camera.
Adds power constraints for the Zoom2 camera sensor (Sony IMX046).
If no constraint was applied then ISP stopped working if Vdd2 went
to OPP2 when streaming 8MPixel at 7.5fps.

Thus Vdd2 needs to be kept at OPP3 while streaming with this sensor.

For every read/write to memory the following throughput constraint
is required:

  3280 x 2464 x 2bpp x 7.5fps = 118387 KByte/sec

Because of a known hardware bug the data must be written into a temp buffer
between the Preview and Resizer modules. Thus the pipeline is setup
as follows:
 Sensor -> CSI2 Rcvr -> CCDC -> Preview -> Temp Mem -> Resizer -> Mem

This means for every pixel there will be three L3 operations:
 1) Preview write to Temp Mem
 2) Resizer read from Temp Mem
 3) Resizer write to final Mem

This then multiplies the thru-put by 3.
So the practical tput is:
  118387 x 3 = 355161 KB/s

This constraint successfully holds the Vdd2 to OPP3.

Signed-off-by: Dominic Curran <dcurran@ti.com>