omap3430sdp: Add OMAP3 power constraints for supported onboard cameras.
Adds power constraints for the 3430sdp camera sensors:
 - Micron/Aptina MT9P012.
 - Omnivision OV3640.

If no constraint was applied then ISP stopped working if Vdd2 went
to OPP2 when streaming max size at max framerate that the size can give.

Thus Vdd2 needs to be kept at OPP3 while streaming with this sensor.

For every read/write to memory the following throughput constraint
is required:

For MT9P012:
  2592 x 1944 x 2Bpp x 11fps = 108257 KByte/s
For OV3640:
  2048 x 1536 x 2Bpp x 7.5fps = 46080 KByte/s

Because of a known hardware bug the data must be written into a temp buffer
between the Preview and Resizer modules. Thus the pipeline is setup
as follows:
 Sensor -> CSI2 Rcvr -> CCDC -> Preview -> Temp Mem -> Resizer -> Mem
         (OV3640 only)

This means for every pixel there will be three L3 operations:
 1) Preview write to Temp Mem
 2) Resizer read from Temp Mem
 3) Resizer write to final Mem

This then multiplies the thru-put by 3.
So the practical tput is:
For MT9P012:
  108256.5 x 3 = 324770 KB/s
For OV3640:
  46080 x 3 = 138240 KB/s

This constraints successfully holds the Vdd2 to OPP3.

Based on patch by Dominic Curran in:
  https://review.source.android.com/10957

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>
CC: Dominic Curran <dcurran@ti.com>