DSPBRIDGE: Get clock handles using clk_get_sys
Replace previous id field used to reference clocks with the same
name by placing the device name associated with them. To use this
scheme the clocks handles need to be requested using clk_get_sys
function.

As McBSP clocks are the only ones used by bridge that have a
device associated with them, the names are hard coded in our
own clock module and clock handles are requested on behalf of
mcbsp driver for the dsp to use.

This change should be kept until a proper solution is found to
request mcbsp clocks.

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>