DSPBRIDGE: Deprecate Unregister Object API
The process of registration adds the path to mpu bridge internal
registry to get the location of the dll for further node creations,
by deregistering it the path and key are deleted from brige registry
making the entry invalid for any other nodes.

This can be an scenario for multiple failures if a rogue application,
decides to deregister DLLs that are still in use.

At registry exit, the entire registry is looped and every data key is freed,
data is nulled afterwards.

Discovered-by: Daniel Charles <dcharles@ti.com>
Discovered-by: Priyesh Bisla <bisla@ti.com>
Discovered-by: Anu Sundararajan <sanuradha@ti.com>
Discovered-by: Venky Viswanathan <venkateshv@ti.com>
Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>