OMAP3 : Fix McSPI RX Timeout
Resending this patch to SPI mailing list as well.

McSPI RX timeout issues are reported on some of the OMAP3 custom boards.

After verifying configuration sequence in TRM, there seems to be issue with
McSPI configuration sequence.

The steps to be followed for both TX and RX:
1. Configure OMAP2_MCSPI_CHCONF0 register with required settings such as
TX/RX mode, word length, force chip select(if required).
2. Set MS bit to 0 in OMAP2_MCSPI_CHCTRL0 register to provide the clock.
3. Enable SPI channel in OMAP2_MCSPI_MODULCTRL

But, the sequence used in current code seems to be wrong. Here it is:
1. Set MS bit to 0 in OMAP2_MCSPI_CHCTRL0 register to provide the clock
(This is done during bootup)
2. Enable SPI channel in OMAP2_MCSPI_MODULCTRL
3. Configure OMAP2_MCSPI_CHCONF0 register with required settings such as
TX/RX mode, word length, force chip select(if required).

After correcting configuration sequence, the timeout seems to be not
reproducible anymore.

Signed-off-by: Manjunatha GK <manjugk@ti.com>