Revert "omap34xxcam: Requeue faulty buffers."
This reverts commit c8fe3518179b1445f24f51dbe73db591bb3554e9.

Reason: OMX camera is having a Mismatch between index and pointer
when a buffer is marked with error for some reason
(for example: CSI2 transmission error).

The hack attempted to retry queueing the buffer, but apparently is not requeueing
the same index/pointer, leading to the unintended mismatch.

TODO: Investigate deeper on why the index could be mismatched when
      doing this DQBUF hack call.

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>