omap: disable GAIA I2C1/I2C4 internal pull-ups
This patch disables GAIA I2C1 and I2C4(SR) internal pull-up, to
use only the external HW resitor >= 470 Ohm for the assured
functionality in HS mode.

While testing the I2C in High Speed mode, it was discovered that
without a proper pull-up resitor, there is data corruption during
multi-byte transfers. RTC(time_set) test case was used for testing.

From the analysis done, it was concluded that ideally we need a pull-up
of 1.6 K Ohm (recomended) or atleast 470 Ohm or greater for assured
performance in HS mode.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Signed-off-by: Allen Pais <allen.pais@sasken.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>