OMAP3 :errata 1.176 - Accesses to DDR stall in SDRC
after a Warm-reset

This patch has the workaround for the errata 1.176.
Accesses to DDR stall in SDRC after a Warm-reset.
It generates a S/W reset and keeps DDR memory content.
This patch enables SDRC self refresh on idle request,
puts SDRC in idle and wait until SDRC goes to idle and then
generates reset.
This would need the u-boot/x-loader workaround changes
as well for the reboot to work correctly.

Signed-off-by: Teerth Reddy <teerth@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>