MUSB: Workaround for Ethernet data alignment issue
This patch is a temporary solution and is under discussion in
linux-usb forum.
Data in the ethernet packet does not start at 32bit aligned address.
Since this data has to be sent using DMA, this data has to start at
32 bit aligned address.
To overcome this issue, this patch moves data by 2 bytes backwards in
ethernet packet assuming ethernet header is not required anymore.

Signed-off-by: Nilkesh Patra <nilkesh.patra@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>