OMAP3630: Clock: add clksel_shift to struct clk
Introduces clksel_shift to struct clk and populates the same for
dpll4_m3_ck, dpll4_m4_ck, dpll4_m2_ck, dpll4_m6, dpll4_m5_ck and dpll3_m3.
The purpose of this change is to make it easy for a generic function to
refresh specific CM_CLKSEL registers on a per-clock basis by first
adding then subtracting the shift value from the register.
This is needed as a work around when new j-type DPLL4 feature is
selected in OMAP3630.

Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Vijaykumar GN <vijaykumar.gn@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>