OMAP3630: Clock: add clock enable function to WA PWRDN prob
This patch implements a workaround for the HS divider problem in
OMAP3630 DPLL3 and DPLL4. Without this workaround the divider value goes
to its reset state after the corresponding PWRDN bit is set from the
CM_CLKEN_PLL register. The correct divider value can be restored by
writing a value to the register different than what is in there, and
then re-writing the desired value.
This Work around is applicable for below HS Dividers.
 1. DPLL3_M3
 2. DPLL4_M2
 3. DPLL4_M3
 4. DPLL4_M4
 5. DPLL4_M5
 6. DPLL4_M6

Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>
Signed-off-by: Vijaykumar GN <vijaykumar.gn@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>