drivers:misc: sources for Init manager module
Texas Instruments BT, FM and GPS combo chips/drivers
make use of a single TTY to communicate with the chip.

Kernel Space Init-Manager works along with User-Mode
Init Manager daemon running to maintain the UART state.
It also is a platform driver with a relevant platform device
in the board-*.c along with the list of BT/FM/GPS chip enable
gpio configuration
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>
Signed-off-by: Janakiram Sistla <x0110377@ti.com>