omap_vout: Write to vrfb context
1. FIFO underflow was caused due to the current configuration of the
V4l2/DSS2 driver. The V4l2 was setting up the VRFB DMA to
write to 0 degree context and DSS2 video DMA to read from the
corresponding rotation angle context.
2. This is now configured accordig to the TRM. The V4l2 wtites to the
VRFB context of desised angle that is geometric representation of the
user rotation(clockwise). DSS2 reads from the 0 degree anlge of
the correspoding context.
3. The VRFB buffer setup has been changed to handle the rotation. The
width and height needs to be flipped when rotation is 90/270 degree as
we are now writing to the corresponding angle of the VRFB context.

Signed-off-by: Kishore Y <kishore.y@ti.com>