OMAP3: PM: Splitting voltage and smartreflex class specific code from smartreflex driver.
This patch does the following
1. Create voltage.c and voltage.h files and move all voltage
processor and voltage controller specific code from smartreflex.c
to these two files.
2. Introduce smartreflex class specific hooks in smartreflex.c. With
this it is ensured that all voltage specific code is removed from
smartreflex.c and smartreflex.h.
3. Introduce smartreflex class3 driver smartreflex-class3.c and
make choosing of the class driver menuconfig based.
4. Remove get_vdd1_opp and get_vdd2_opp from smartreflex.c. This is
now needed by voltage.c and smartreflex-class3.c. To keep the implementaion
clean these functions are exported out in resource34xx.c. But this
should be removed once the new opp framework is accepted.
5. Disable the smartreflex across both frequency and voltage scaling
instead of just across voltage scaling as before. This is a h/w
recommended practice. This step is required in this patch. Otherwise
with voltage functions separated out, dvfs with smartreflex will
fail.

Signed-off-by: Thara Gopinath <thara@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>