pm: opp: twl/tps: optimize uv to vsel function
For integer values x and y; int div x / y causes truncation. Current
omap_twl_uv_to_vsel function implements an equivalent of ceil which
is based on an if condition to check truncation and round up. We can
do this in a more optimal manner without the if condition. The round
up is handled by adding the round off factor prior to truncation as:
(x + (y - 1)) / y

Cc: Kevin Hilman <khilman@deeprootsystems.com>
Acked-by: Nishanth Menon <nm@ti.com>

Signed-off-by: Romit Dasgupta <romit@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>