OMAP3 PM: Conditional UART context save restore.
Currently UART context save is done in idle
thread thru a call to omap_uart_prepare_idle irrespective of
what power state is attained by the power domain to which
the UART belongs to. This patch allows omap_uart_prepare_idle
to take power state as a parameter and this function in
turn does a uart context save only if the passed power state
is PWRDM_POWER_OFF. In the restore path a restore will
happen only if a valid save has happened.

Signed-off-by: Thara Gopinath <thara@ti.com>
Signed-off-by: x0124779@ti.com
Signed-off-by: Vijaykumar GN <vijaykumar.gn@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>