OMAP3 PM: Adding conditional core powerdomain context save and restore
This patch adds a parameter core_state , depicting the power state
which the core domain will attempt to enter, to omap3_core_save_context
and omap3_core_restore_context. This is so as to distinguish between
the context save and restore required when core domain is attempting
OSWR and OFF. Core OSWR does not require interrupt controller,
system control module and dma controller context save and retore
where as Core OFF reuires these.

Signed-off-by: Thara Gopinath <thara@ti.com>
Signed-off-by: x0124779@ti.com
Signed-off-by: Vijaykumar GN <vijaykumar.gn@ti.com>
Signed-off-by: Manjunatha GK <manjugk@ti.com>