OMAP: DSS: Fix for OFFmode for HDMI
When OFF mode is enabled,Core is getting switched off even
when clock's of dss are still in use.
Since core is going to oFF mode gpio pins are being reset.
Hence display goes blank.With the dma latency constrains
we are making the core not go to OFF mode until dss clock's
are in use.

Signed-off-by: Sudeep Basavaraj <sudeep.basavaraj@ti.com>