DSPBRIDGE: Avoid NULL pointer dereference
Pointers are checked before they are dereferenced.

Also, in function DMM_GetHandle, the following code is added in order to
ensure that phDmmMgr is NULL when failed to get the Dev Object handle:

else
	*phDmmMgr = NULL;

So that it is easier to check if pointer is NULL after DMM_GetHandle
function is called.

Signed-off-by: Ivan Gomez <ivan.gomez@ti.com>